requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
psycopg2-binary==2.9.1
pyTelegramBotAPI==3.8.2
requests==2.26.0
schedule==1.1.0
spbuTimetableAPI==1.0.3
urllib3==1.26.6
